Rename module & package name to stream-log-android-file
