+++ app/build.gradle.kts
@@ -36,7 +36,7 @@ dependencies {
     implementation(project(":stream-log"))
     implementation(project(":stream-log-file"))
     implementation(project(":stream-log-android"))
-    implementation(project(":stream-log-android-init"))
+    implementation(project(":stream-log-android-file"))
 
     // compose
     implementation(Dependencies.composeUi)
