+++ settings.gradle
@@ -17,4 +17,4 @@ include ':app'
 include ':stream-log'
 include ':stream-log-file'
 include ':stream-log-android'
-include ':stream-log-android-init'
+include ':stream-log-android-file'
