+++ stream-log-android-file/api/stream-log-android-file.api
@@ -1,31 +1,31 @@
-public final class io/getstream/log/android/init/BuildConfig {
+public final class io/getstream/log/android/file/BuildConfig {
 	public static final field BUILD_TYPE Ljava/lang/String;
 	public static final field DEBUG Z
 	public static final field LIBRARY_PACKAGE_NAME Ljava/lang/String;
 	public fun <init> ()V
 }
 
-public final class io/getstream/log/android/init/StreamLogFileManager {
-	public static final field INSTANCE Lio/getstream/log/android/init/StreamLogFileManager;
+public final class io/getstream/log/android/file/StreamLogFileManager {
+	public static final field INSTANCE Lio/getstream/log/android/file/StreamLogFileManager;
 	public final fun clear ()V
-	public final fun init (Lio/getstream/log/android/init/StreamLogFileManager$ShareManager;Lio/getstream/log/android/init/StreamLogFileManager$ClearManager;)V
+	public final fun init (Lio/getstream/log/android/file/StreamLogFileManager$ShareManager;Lio/getstream/log/android/file/StreamLogFileManager$ClearManager;)V
 	public final fun share ()V
 }
 
-public abstract interface class io/getstream/log/android/init/StreamLogFileManager$ClearManager {
+public abstract interface class io/getstream/log/android/file/StreamLogFileManager$ClearManager {
 	public abstract fun clear ()V
 }
 
-public abstract interface class io/getstream/log/android/init/StreamLogFileManager$ShareManager {
+public abstract interface class io/getstream/log/android/file/StreamLogFileManager$ShareManager {
 	public abstract fun share ()V
 }
 
-public final class io/getstream/log/android/init/StreamLogFileProvider : androidx/core/content/FileProvider {
+public final class io/getstream/log/android/file/StreamLogFileProvider : androidx/core/content/FileProvider {
 	public fun <init> ()V
 	public fun onCreate ()Z
 }
 
-public final class io/getstream/log/android/init/StreamLogFileService : android/app/Service {
+public final class io/getstream/log/android/file/StreamLogFileService : android/app/Service {
 	public fun <init> ()V
 	public fun onBind (Landroid/content/Intent;)Landroid/os/IBinder;
 	public fun onStartCommand (Landroid/content/Intent;II)I
