+++ stream-log-android-file/build.gradle
@@ -7,7 +7,7 @@ apply plugin: 'de.mannodermaus.android-junit5'
 
 ext {
     PUBLISH_GROUP_ID = Configuration.artifactGroup
-    PUBLISH_ARTIFACT_ID = 'stream-log-android-init'
+    PUBLISH_ARTIFACT_ID = 'stream-log-android-file'
     PUBLISH_VERSION = rootVersionName
 }
