+++ stream-log-android-file/src/main/AndroidManifest.xml
@@ -15,13 +15,13 @@
      limitations under the License.
 -->
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="io.getstream.log.android.init"
+    package="io.getstream.log.android.file"
     >
 
     <application>
 
         <service
-            android:name="io.getstream.log.android.init.StreamLogFileService"
+            android:name=".StreamLogFileService"
             android:enabled="true"
             android:exported="true"
             >
@@ -37,7 +37,7 @@
         </service>
 
         <provider
-            android:name="io.getstream.log.android.init.StreamLogFileProvider"
+            android:name=".StreamLogFileProvider"
             android:authorities="${applicationId}.streamlogfileprovider"
             android:exported="false"
             android:grantUriPermissions="true"
