+++ stream-log-android-file/src/main/res/xml/log_file_paths.xml
